# api/docs/swagger.yaml
definitions:
...
  auth_service.AddUserRelationRequest:
    properties:
      relation_id:
        type: string
      user_id:
        type: string
    type: object
...
  auth_service.CreateUserRequest:
    properties:
      active:
        type: integer
      client_platform_id:
        type: string
      client_type_id:
        type: string
      email:
        type: string
      expires_at:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      photo_url:
        type: string
      project_id:
        type: string
      role_id:
        type: string
    type: object
...
  auth_service.GetUserListResponse:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/auth_service.User'
        type: array
    type: object
...
  auth_service.UpdateUserRequest:
    properties:
      active:
        type: integer
      client_platform_id:
        type: string
      client_type_id:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      login:
        type: string
      name:
        type: string
      phone:
        type: string
      photo_url:
        type: string
      project_id:
        type: string
      role_id:
        type: string
    type: object
...
  auth_service.User:
    properties:
      active:
        type: integer
      client_platform_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      photo_url:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
    type: object
...
  auth_service.UserRelation:
    properties:
      relation_id:
        type: string
      user_id:
        type: string
    type: object
  auth_service.UserRelationPrimaryKey:
    properties:
      relation_id:
        type: string
      user_id:
        type: string
    type: object
...
  http.Response:
    properties:
      data:
        type: object
      description:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  license: {}
  termsOfService: https://udevs.io
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Get User List
      operationId: get_user_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetUserListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetUserListResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get User List
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create User
      operationId: create_user
      parameters:
      - description: CreateUserRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update User
      operationId: update_user
      parameters:
      - description: UpdateUserRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update User
      tags:
      - User
...
  /user-relation:
    delete:
      consumes:
      - application/json
      description: Get UserRelation
      operationId: delete_user_relation
      parameters:
      - description: UserRelationPrimaryKeyBody
        in: body
        name: user-relation
        required: true
        schema:
          $ref: '#/definitions/auth_service.UserRelationPrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete UserRelation
      tags:
      - UserRelation
    post:
      consumes:
      - application/json
      description: Create UserRelation
      operationId: add_user_relation
      parameters:
      - description: AddUserRelationRequestBody
        in: body
        name: user-relation
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddUserRelationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: UserRelation data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.UserRelation'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create UserRelation
      tags:
      - UserRelation
  /user/{user-id}:
    delete:
      consumes:
      - application/json
      description: Get User
      operationId: delete_user
      parameters:
      - description: user-id
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete User
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get User By ID
      operationId: get_user_by_id
      parameters:
      - description: user-id
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UserBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get User By ID
      tags:
      - User
swagger: "2.0"